LayoutTests/http/tests/security/contentSecurityPolicy/worker-importscripts-blocked-expected.txt

PASS result is "importScripts blocked: NetworkError (DOM Exception 19):  A network error occurred."
PASS successfullyParsed is true

TEST COMPLETE
